requirements.txt
beautifulsoup4==4.9.3
grpcio==1.36.1
numpy==1.20.1
pandas==1.2.3
protobuf==3.15.6
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
soupsieve==2.2
